fix: cast timestep embedding to model dtype in PredictorP
get_timestep_embedding returns fp32 (Python literals + default torch.arange
dtype). During training, accelerator's mixed_precision autocast bridged
this to bf16 for the time_proj matmul. At inference there's no autocast
context, so the fp32 input hits fp16/bf16 Linear weights and errors.

Cast explicitly before time_proj so the model works in any dtype context
without needing a wrapping autocast.

+++ src/latte_physics.py
@@ -141,8 +141,12 @@ class PredictorP(nn.Module):
         # 2. Project text embeddings
         text_features = self.text_proj(text_embeddings)  # [B, text_seq_len, hidden_dim]
         
-        # 3. Get timestep embeddings
+        # 3. Get timestep embeddings — cast to model dtype so it matches
+        # time_proj's weight precision (fp16/bf16). Training relied on
+        # accelerator's autocast to bridge fp32 -> low-precision matmul; at
+        # inference there's no autocast context so we cast explicitly.
         time_emb = self.get_timestep_embedding(timesteps, self.hidden_dim)  # [B, hidden_dim]
+        time_emb = time_emb.to(dtype=self.time_proj[0].weight.dtype)
         time_emb = self.time_proj(time_emb).unsqueeze(1)  # [B, 1, hidden_dim]
         
         # 4. Fuse: latent + text + time
